+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+* use `vite` and simplify the project configuration
+* removed the unused `ICursorPosObserver` interface and related code
+* removed the unused `onDestroy` parameter to `MathView` 
+
 ## [Releases]
 
 ### [0.2.2] - 2021-06-24